requirements.txt
absl-py==2.3.1
affine==2.4.0
aiobotocore==2.24.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
anndata==0.11.4
array-api-compat==1.12.0
asciitree==0.3.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.5
botocore==1.40.18
cellcharter==0.3.5
certifi==2025.8.3
charset-normalizer==3.4.3
chex==0.1.90
click==8.3.0
click-plugins==1.1.1.2
cligj==0.7.2
cloudpickle==3.1.1
colorama==0.4.6
colorcet==3.1.0
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
dask==2024.11.2
dask-expr==1.1.19
dask-image==2024.5.3
datashader==0.18.2
debugpy==1.8.17
decorator==5.2.1
docrep==0.3.2
etils==1.13.0
exceptiongroup==1.3.0
executing==2.2.1
fasteners==0.20
filelock==3.19.1
flax==0.10.7
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
gdown==5.2.0
geopandas==1.1.1
grpcio==1.75.0
gseapy==1.1.10
h5py==3.14.0
hotspotsc==1.1.3
humanize==4.13.0
idna==3.10
igraph==0.11.9
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflect==7.5.0
ipykernel==6.30.1
ipython==8.37.0
jax==0.6.2
jaxlib==0.6.2
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lazy_loader==0.4
legacy-api-wrap==1.4.1
leidenalg==0.10.2
lightning==2.5.5
lightning-utilities==0.15.2
llvmlite==0.45.0
locket==1.0.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
matplotlib-scalebar==0.9.0
mdurl==0.1.2
ml_collections==1.1.0
ml_dtypes==0.5.3
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
mudata==0.3.2
multidict==6.6.4
multipledispatch==1.0.0
multiscale_spatial_image==2.0.3
muon==0.1.7
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.4.2
newick==1.0.0
numba==0.62.0
numcodecs==0.13.1
numpy==2.2.6
numpyro==0.19.0
ome-zarr==0.11.1
omnipath==1.0.12
opencv-python==4.12.0.88
opt_einsum==3.4.0
optax==0.2.6
orbax-checkpoint==0.11.25
packaging==25.0
pandas==2.3.2
param==2.2.1
parso==0.8.5
partd==1.4.2
patsy==1.0.1
pillow==11.3.0
PIMS==0.7
platformdirs==4.4.0
pooch==1.8.2
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
pure_eval==0.2.3
pyarrow==21.0.0
pyct==0.5.0
Pygments==2.19.2
pynndescent==0.5.13
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.7.1
pyro-api==0.1.2
pyro-ppl==1.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.5
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.1.0
rasterio==1.4.3
requests==2.32.5
rich==14.1.0
s3fs==2025.9.0
scanpy==1.11.4
scArches==0.6.1
#scHPL==1.0.5
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
scvi-tools==1.3.3
seaborn==0.13.2
session-info2==0.2.2
session_info==1.0.1
shapely==2.1.2
simplejson==3.20.1
six==1.17.0
sknw==0.15
slicerator==1.1.0
soupsieve==2.8
sparse==0.17.0
spatial_image==1.2.3
spatialdata==0.5.0
spatialdata-plot==0.2.12
squidpy==1.6.5
stack-data==0.6.3
statsmodels==0.14.5
stdlib-list==0.11.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorstore==0.1.77
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.5.10
toolz==1.0.0
torch==2.8.0
torchaudio==2.8.0
torchgmm==0.1.3
torchmetrics==1.8.2
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
treescope==0.1.10
typeguard==4.4.4
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
validators==0.35.0
wcwidth==0.2.14
Werkzeug==3.1.3
wrapt==1.17.3
xarray==2025.6.1
xarray-dataclass==3.0.0
xarray-schema==0.0.3
xarray-spatial==0.4.0
yarl==1.20.1
zarr==2.18.3
zipp==3.23.0
